fixed some review suggestions
