+++ client/src/components/Header.js
@@ -1,11 +1,11 @@
 import styled from "styled-components";
 
-import { ReactComponent as fin2u } from "../images/fin2uLogo.svg";
+import { ReactComponent as Fin2u } from "../images/fin2uLogo.svg";
 
 export default function Header() {
   return (
     <Container>
-      <Logo />
+      <Fin2u />
     </Container>
   );
 }
@@ -14,10 +14,10 @@ const Container = styled.div`
   display: flex;
   flex-direction: column;
   align-items: flex-end;
-`;
 
-const Logo = styled(fin2u)`
-  width: 49.87px;
-  height: 23.17px;
-  margin: 0.5rem;
+  svg {
+    width: 49.87px;
+    height: 23.17px;
+    margin: 0.5rem;
+  }
 `;
