+++ client/src/pages/Start.js
@@ -18,7 +18,7 @@ export default function Start({
           userToCalculate={userToCalculate}
         />
       )}
-      {componentToDisplay === "value" && (
+      {componentToDisplay === "solution" && (
         <DisplayInsurances
           insurances={insurances}
           setComponentToDisplay={setComponentToDisplay}
