+++ server/src/models/user.models.js
@@ -1,15 +1,42 @@
 import mongoose from "mongoose";
 
 const userSchema = {
-  name: String,
-  lastname: String,
-  age: Number,
-  birthday: String,
-  relationship: String,
-  children: Number,
-  jobStatus: String,
-  income: Number,
-  netIncome: Number,
+  name: {
+    type: String,
+    required: true,
+  },
+  lastname: {
+    type: String,
+    required: true,
+  },
+  age: {
+    type: Number,
+    required: true,
+  },
+  birthday: {
+    type: String,
+    required: true,
+  },
+  relationship: {
+    type: String,
+    required: true,
+  },
+  children: {
+    type: String,
+    required: true,
+  },
+  jobStatus: {
+    type: String,
+    required: true,
+  },
+  income: {
+    type: String,
+    required: true,
+  },
+  netIncome: {
+    type: String,
+    required: true,
+  },
   houseOwner: Boolean,
   valuables: Boolean,
   car: Boolean,
